requirements.txt
asgiref==3.5.2
coverage==6.5.0
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1.1
django-adminlte2==0.4.3
django-stdimage==6.0.1
gunicorn==20.1.0
model-mommy==2.0.0
Pillow==9.2.0
psycopg2-binary==2.9.3
sqlparse==0.4.3
static3==0.7.0
tzdata==2022.4
